Discord2Discord/requirements.txt
discum==1.1.0
requests>=2.31.0
discord.py>=2.3.2
python-dotenv>=1.0.1
aiohttp>=3.9.5